refactor updatejob to allow video upload
Refactor the existing genupdatejob and applyupdatejob to be more
generic.

applyupdatejob is refactored to allow both creation and updating of
jwplatform resources.

genupdatejob is refactored to examine the existing jwplatform items and
generate a list of create/update jobs from them. We take the opportunity
to only upload one clip from each media item (preferring video).