introduced 2 new update jobs:
'image_load': The job is generated when SMS media has a thumbnail and the
image_md5 params don't match. When run, the job uploads the SMS thumbnail
and sets an image_status=loaded custom param.

'image_check': The job is generated when SMS media has a thumbnail and the
image_status custom param is 'loaded'. When run, the job checks JWPlatform
for the creation status of the thumbnail as sets on the image_status param.