widen range of acceptable export formats
Widen the range of formats we try to upload to JWP. Some media items had
neither an audio or video archive export.

Modify the CSV dump script to dump *all* clips and define a preferred
ordering for formats, video then audio and within each decreasing
quality options. Explicitly log items with no media at all (there are
some!) or ones where we could not determine a "best" format.

IMPORTANT: This changes the default strip-leading path option from 1 to
0 to allow for non-archive media items to be downloaded. This needs to
be combined with a change to gen_jwplayer_feed.sh to strip the
"sms-archive02/" from the base URL.