genupdatejob: add property to collection listing media ids
Add a property to the collection listing all the media ids for the
collection in the order they should be in. This custom property is not
modified by adding/removing videos to the corresponding JWP playlist
unlike sms_{,failed_}media_ids and is used by the media platform to
reconcile videos with playlists.